requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.21
matplotlib>=3.5
pandas>=1.5
PyYAML>=6.0
timm>=0.9.12
segmentation-models-pytorch>=0.3.3source .venv/bin/activate
segyio>=1.9.0